Make Provider middleware idempotent
Simulation tests will currently call middleware functions several times in a row. This middleware fails due to changing originalMiddleware, and is not idempotent. This should allow us to call this middleware several times with the same result.

The issue is seen in the simulation tests added here: https://github.com/fusionjs/fusion-cli/pull/302