Pass query string after stripping route prefix
When stripping out the route prefix, the only changes strip-prefix made
was making sure the route had a `/` if the prefix was the only part of
the url. This change will make sure that the `/` gets added back if the
url has a prefix followed directly by a query string. Before it could
cause the ssr plugin to throw an error because ctx.path would be null.