Add support for `targetRuleName` in manifest
Jazelle provides the `web_library` rule for defining a target and its deps. When auto-updating the deps in `BUILD.bazel` based on `package.json` changes, Jazelle assumes the build file is using `web_library` directly, instead of via a macro, for example.

This adds support for a `targetRuleName` property in `manifest.json` to provide an alternate rule name to reference when auto-updating the deps