Kill the old node process upon reload when debugging in development
I'm trying to use VSCode debugger auto-attach in conjunction with `fusion dev --debug`; it works well on the first run, but I'll get an error along the lines of `Starting inspector on 127.0.0.1:9229 failed: address already in use` whenever I make changes and the process reloads. It seems like the old node process doesn't reliably get killed before the new one is spawned and attempts to attach to the debug port.

This PR does the job but may be a bit heavy-handed; if there's a better approach I'd love to hear it (I tried sending different kill signals in `dev-runtime#killProc` without any success).

Co-authored-by: Chris Deacy <deacy@uber.com>