Allow serving source maps in a staging environment
Enable serving source maps based on a runtime environment variable.

### Build
This PR introduces additional build artifacts for each JS bundle
```diff
.fusion/dist/production/client/
  - client-main-xxx.js (still contains no source map comment)
  - client-main-xxx.js.br
  - client-main-xxx.js.gzip
  - client-main-xxx.js.map
+ - client-main-xxx-with-map.js (contains source map comment)
+ - client-main-xxx-with-map.js.br
+ - client-main-xxx-with-map.js.gzip
```

*The only difference between `client-main.js` and `client-main-with-map.js` is the addition of the source map comment at the end of the file, i.e.:*
```js
//# sourceMappingURL=client-main.js.map
```

This is technically the wrong file name, i.e. (`client-main-with-map.js` source map comment is `client-main.js.map` without `-with-map`), but this doesn't actually cause any issues since the bundles are identical.

The "cost" of writing files twice is mostly negligible. There is no change in build time and the JS weight of a production build is a minority of the final build size when using lots of images/svgs/fonts. E.g. in a particularly big Fusion app:

```bash
$ du -sh prod-build*
233M    prod-build/
242M    prod-build-with-duplicates/
```

### Runtime
The SSR template will use use `-with-map.js` bundles when `DANGEROUSLY_EXPOSE_SOURCE_MAPS` is true, whether `CDN_URL` is set or not.

`CDN_URL` will be ignored (not used) when source maps are exposed to prevent caching.

The runtime chunk is rewritten to reference `-with-map` async chunks in the webpack bootstrap code. i.e.
```js
function jsonpScriptSrc(chunkId) {
  return (
    __webpack_require__.p +
    "client-" +
    ({"0": "renamed-chunk"}[chunkId] || chunkId) +
    "-" +
    {"0": "99e217316a4378348969-with-map", "3": "bef335caaff3e2956271-with-map"}[chunkId] +
    ".js"
  );
}
```
Unfortunately, this is just find/replace, but it has limited impact since it only affects the runtime bundle.

#### Why `DANGEROUSLY`?

If you expose source maps on a publicly accessible server, or behind a CDN, your source code will be public. `CDN_URL` is ignored when source maps are exposed to prevent caching but they will still be exposed. Using `DANGEROUSLY` will bring attention to potential security concerns, and hopefully prompt users to read and understand the drawbacks. There isn't a complete solution to this without being too heavy-handed or Uber-specific, so it's up to the user to consider the concerns.

Co-authored-by: Mickey Burks <mburks@uber.com>