Make install always generate same lockfiles
Installing from a service searches upstream then downstream again, but installing from a library does not search downstream first as this would require recursively zigzagging up and down to find all nodes in the graph. This means that depending on where an installation is triggered from, it's possible to miss a resolution, resulting in different lockfiles if installation is triggered from different projects.

Instead, this change checks all lockfiles all the time