Update index.js
Flowtype rules shouldn't be included in the Fusion eslint rules because Fusion can be run on systems that don't include flowtype, such as Typescript. Requiring these rules would force projects to install eslint-plugin-flowtype even though they don't need it to run Fusion. Those projects would then have to include the following to disable flowtype linting errors:

```
rules: {
    "flowtype/require-valid-file-annotation": 0,
    "flowtype/no-types-missing-file-annotation": 0,
```

A recommendation is to split flowtype and typescript rules into their own packages, such as eslint-config-fusion-flowtype and eslint-config-fusion-typescript. The users of Fusion can choose which package to install, or the scaffolding tool can select the appropriate package depending on the type of project (if that choice becomes available in the future).