Fix import for non-public `ReactReduxContext`
This import is technically not a part of the public API, so it doesn't appear in the libdef. The libdef wasn't installed, so this wasn't caught in the last release.

While `ReactReduxContext` is not a public API, this plugin is meant to wrap `react-redux` and therefore should be closely tied to the `react-redux` version (if the API changes, we will update this plugin).

Co-authored-by: Mickey Burks <mburks@uber.com>