Do align on add
Previously, `jazelle add` would add the latest version of a package, regardless of what versionPolicy said. This meant that if other projects in the monorepo used a version of a dep that was out of date but tracked by the version policy, adding that dep to a project would put the repo in a bad state.

This change makes it so that if a dep is tracked by the version policy, `jazelle add` will add that version by default. This can be overridden by explicitly specifying a version (e.g. `jazelle add foo@^1.0.0`)