Make `jazelle install` use conservative lockfile updating algorithm
Previously, by default, if `jazelle install` detected that transitive version range was not present in a lockfile, it would pull from the NPM registry, regardless of whether compatible versions existed in the lockfile. It would then proceed to upgrade all other lockfiles to match. Now it will prefer to collapse to a compatible version in the lockfile, with the end effect of leaving other lockfiles alone

`jazelle upgrade` and `jazelle add` still update transitives to the NPM version