Add support for basic wildcard matching in the dist attr
This adds support for basic wildcard matching in the dist attr of our bazel macros. It only works with a single wildcard expansion, and will throw an error otherwise. I purposely didn't make this feature complete because I think we need to revisit our API here to try and handle the cases for dist and gen_srcs in a unified manner.