Match jazelle script command exit code to subcommand exit code
Before this change calling `jazelle [package.json script]` with a non 1 exit code would always return 1. This PR propagates that exit code through the cli. This is currently done by decorating the Error object with a status code, but I could imagine creating a specific Error class to handle it if preferred.