Add error document references
This adds some developer-friendliness around DI errors within `fusion-core`. A `DIError` can now reference a markdown document that further describes the error and how to mitigate. A link to the document will be visible on the server-error page. Also, the error stack traces will now point directly to the problematic user code.

The next bit of work will be to update the server-error page, potentially showing code location from the stack trace.

Co-authored-by: Mickey Burks <mburks@uber.com>