Ensure -with-map asset is produced for chunks that only contain JSON modules
There's a regression after webpack 5 upgrade, where `-with-map` asset is not being generated for chunks that only contain JSON modules. Webpack does not produce a source map for such chunks / modules, hence we need to handle it properly to ensure the `-with-map` asset is produced for every JS asset (i.e. even in cases when source map is not available).

Co-authored-by: Siarhei Fedarovich <siarhei@uber.com>