Ensure split chunk ids stay in sync on both client and server
This ensures that ImportDependencyTemplate is applied using fresh client chunk ids on the server. Previously it'd not be updated, as module's hash would not reflect the change (in which case webpack skips code generation and uses previously cached result).

Also fixes an issue, where MergedDeferredState with ClientChunkIndex was not reset properly in dev mode.