fix(fusion-cli): disable webpack 5 automatic publicPath
Webpack 5 comes with automatic publicPath resolution by default, which is not well supported in older browsers, and is causing runtime errors in IE11 for example.

Fusion.js has custom logic to determine the publicPath that can be controlled via ENV variables at runtime, hence we're opting-out of this new webpack 5 default behavior.

Co-authored-by: shYkiSto <siarhei@uber.com>