Add experimental esbuild minifier
Introducing `experimentalEsbuildMinifier` compiler option to replace Terser minifier with esbuild. By doing so we've seen the production build times go down by ~30%. At the same time esbuild is not an optimal choice yet, as it does not support the same level of minification as Terser: https://github.com/evanw/esbuild/issues/639#issuecomment-792057348, which results in larger bundle sizes (+ ~10%). This is something web developers will need to take into account when choosing to opt-in to use esbuild miinifer to achieve faster production builds.