Use in-order scheduling policy for babel-worker farm
We use jest-worker farm to transpile modules with babel, and to start a new worker we are paying a penalty of loading the babel configuration, which in turn loads a bunch of modules (presets, plugins), and all of this takes ~300-500ms for each worker. Using `in-order` scheduling policy allows us to gradually load new workers as we need them to run more tasks in parallel, this is different from default `round-robin` load balancing, such that we'll gradually load new workers in the farm in case all existing workers are busy, which allows us to schedule more serial tasks on hot workers. [See more](https://github.com/facebook/jest/pull/10902)