Add support to Beacon API in fusion-plugin-universal-events
Use [Beacon API](https://developer.mozilla.org/en-US/docs/Web/API/Beacon_API) instead of [Fetch API](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API) as transport to send events in `fusion-plugin-universal-events`.
Keeping `fetch` as a fallback if the browser does not support Beacon API or `navigator.sendBeacon` call fails.

### Why?
The plugin batches events together before sending them and batches are often sent on page closing. Most of the `fetch` requests are cancelled if the page is closing so this causes a lot of events to be lost due to cancelled TCP connections. The `navigator.sendBeacon` is meant to be used for this very purpose: it sends `POST` requests with low priority and it delivers them even if the page is closed. Also, the low priority lets the app properly load the UI first, leading to a smoother experience for the user.

### Prior work

The Beacon API as a transport system for events has already been tested on the _web-payments-experience_ service. The experiment was split into different phases:

1. We sent a test event at app startup both with the Beacon and the Fetch APIs - we received almost 100% of the events sent with the Beacon API and about 30% of the events sent with the old plugin
2. We selected an actual production event and made half of the users send it via Beacon API and the other half via the old plugin - we received 60% of beacons and 40% of XHR requests.
3. We sent all the events via Beacon API - the total number of events increased by ~70%.

For more info about the experiment results, please take a look [here](https://docs.google.com/document/d/1OSYerFBtgU2og7BhcgwyE_yMeXnFVthcTrDzQZBRJDA/edit#).

Co-authored-by: enrico-uber <78430438+enrico-uber@users.noreply.github.com>
Co-authored-by: Ryan Tsao <ryan.j.tsao@gmail.com>
Co-authored-by: shYkiSto <siarhei@uber.com>