Add more event handlers for page leave event in fusion-plugin-universal-events
`visibilitychange` event does not work well on some browsers, so add event handlers for `pagehide`
and `beforeunload` events too. Do not add event handler for `unload` event because on some browsers
it disables `back/forward-caching`.
Add event listener for `visibilitychange` on `document` as `EventTarget` instead of `window` because
it works better on some old versions of Safari.