Change default event persistence from localStorage to in-memory
Because `navigator.sendBeacon` is now used, I think sending of events should be more reliable. One problem with `localStorage` batching is persisted events can intermix with other Fusion apps on the same subdomain. This is a bit of a footgun and I think with `navigator.sendBeacon`, `localStorage` persistence shouldn't be necessary.

See https://t3.uberinternal.com/browse/WPT-7960 for more context.

Alternatively, we could change how `fusion-plugin-events-adapter` works. Currently it injects `SVC_ID` and `UBER_RUNTIME_ENVIRONMENT` on the server to any batched events (which could have been persisted from any other deployment on the same domain). Instead, we'd need to send down  `SVC_ID` and `UBER_RUNTIME_ENVIRONMENT` to the client so they are included at the time they are persisted.

An open question is how to handle the case where client-side event from service A is persisted but sent to service B. Which  `SVC_ID` and `UBER_RUNTIME_ENVIRONMENT` should be used? The one from the service A or service B? I would assume service A would take precedence, but it seems a bit murky. If we just avoid persistence altogether this simply won't occur.

Co-authored-by: Ryan Tsao <ryan.j.tsao@gmail.com>