Ensure i18n manifest is populated using chunk's modules
This fixes an issue where i18n manifest could be populated with modules that were not included by any of the client chunks, which lead to a build error when compiling server entry, which expects all modules listed in i18n manifest to have a corresponding client chunk id. The modules can be excluded from the bundle for valid reasons (e.g. as a result of optimization). Both client chunk metadata and i18n manifest will be built from the same source (compilation.chunks and corresponding compilation.chunkGraph).