Avoid sending payloads larger than max xhr limit
The existing max payload logic depends on actually sending the payload and receiving a 413 response. This puts more load on the server, particularly in the case of really large payloads, where several retry attempts might be made (up to 9, I suppose).

Instead, we should limit the size to the known max body limit (1mb) of the body parser before even attempting to send the payload in the first place.

This PR preserves the existing backoff logic, but also adds logic to avoid sending payloads larger than 1mb in the first place. This should cut down on retries and thus reduce server load.