Add file support for formData objects
Currently, `fusion-plugin-rpc` is able to support key/value pairs that are stored in formData objects, but does not currently process file objects that can be sent through formData. This PR intends to add support for parsed files from formData without breaking the existing implementation for parsing formData key/value pairs.

With this new addition, files will be assigned a key based on the name assigned to them when initially appended to a formData object, and exist in the resulting key/value pairs that will be stored in the `body` object that is returned by `fusion-plugin-rpc`.

Co-authored-by: ghostfruitleaf <49515396+ghostfruitleaf@users.noreply.github.com>
Co-authored-by: shYkiSto <siarhei@uber.com>