Do not attempt to run module before all of its deps are resolved
There's a race condition w/ a `split()` component, where it prematurely requires a module once it gets populated in the `__webpack_modules__` global registry, i.e. whenever the chunk containing the module is loaded. This does not work for cases, where this module depends on other module from another async chunk (which may not have finished loading by the time we attempt to require it), or another case where translations may not be ready, causing placeholder strings to be used in place of translated strings on the UI. This usually leads to abrupt / broken UX.

I'm removing the fore spoken optimization, so that we keep the right order of things, where we render the deferred component when all of its dependencies get resolved.

Co-authored-by: shYkiSto <siarhei@uber.com>