Verify all chunks are loaded before requiring async module
There's a race condition w/ a split() component, where it prematurely requires a module once it gets populated in the `__webpack_modules__`  global registry, i.e. whenever the chunk containing the module is loaded. This does not work for cases, where this module depends on other module from another async chunk (which may not have finished loading by the time we attempt to require it), or another case where translations may not be ready, causing placeholder strings to be used in place of translated strings on the UI. This usually leads to abrupt / broken UX.

I had first attempted to remove the fore spoken optimization, so that we keep the right order of things, where we render the deferred component when all of its dependencies get resolved. But apparently, there are apps that skip `prepare()` on the client, while it still runs on the server. In which case the AsyncComponent is not populated before the page is hydrated. Causing a rendering mismatch.

This ensures all dependent chunks are loaded, before the async module is synchronously required by the split component.

Co-authored-by: shYkiSto <siarhei@uber.com>