Using fetch in case a single event is larger than the beacon limit
When a large payload is to be sent ( > 60k chars ) using sendBeacon, the queuing logic currently ends up sending an empty payload and this large payload remains in the pending list. This keeps repeating for every subsequent flush and all the payloads after this one are never sent.

To fix this, we can use fetch to send the event in case a single event is larger than the beacon limit ( 60k chars ).