For the connected router plugin, pin fusion-plugin-react-router to 2.3.3
This PR is in preparation of upgrading `react-router` inside of `fusion-plugin-react-router` to 6.x.

`connected-router` does not support react-router 6 and going by the github repo commits and issues, they have no plans to. See https://github.com/supasate/connected-react-router/issues/543. Seems like community members are migrating to other solutions.

In lieu of this, this PR pins this plugin to the latest version of `fusion-plugin-react-router` to serve as a time capsule for whomever still wants to use this. That way when we upgrade `fusion-plugin-react-router` it won't break this plugin.

Alternatively, there is potential work to migrate this plugin to use something else under the hood that supports react router 6 but to me that should be an entirely new plugin instead of continuing to work on this one.