Remove dependency on enzyme from tests
Remove dependency on enzyme from tests
* Use `@testing-library/react` where possible
* Removed obsolete react@15 jest setup from fusion-cli
* Moved `enzyme` to optional `peerDependencies` for fusion-cli package

This is done in preparation for react@17, with future goal to not include any of the enzyme-craft in the fusion-cli out of the box. There's no official adapter for react@17 yet([and very likely that there won't be any](https://dev.to/wojtekmaj/enzyme-is-dead-now-what-ekl)), and same story continues for react@18. We may use unofficial `@wojtekmaj/enzyme-adapter-react-17` for the time being for the react@17 upgrade, to give users more time to refactor their enzyme tests to use different renderer / utils.