Add test for middleware order
Ensure that middleware for enhancers are always executed in the same position as the original token, even if registered much later. We had no test coverage for this emergent behavior of our DI algorithm.

In practice, implicit (non-token) order dependencies for middleware execution is common. So it is crucial that registration-dependent middleware order execution is always consistent, otherwise apps can break. This PR adds test coverage for this kind of behavior to so middleware order doesn't accidentally change in the future.