Fix location passed to server router
The `location` prop passed to the base Router object from React Router should include any query params that are present in the URL. Right now the implementation is just passing the path which causes the extra URL params to be stripped when using context hooks like `useLocation`.

Also adds a test to verify behavior.