Bundle packages that could not be resolved at build time
Update webpack externals configuration, so that we fallback to webpack to handle the dependency in case its absolute path could not be resolved using the node.js resolver at build time (as opposed to assuming that the dependency may be available later at runtime). This also ensures a failing build in case the requested module is missing, which in turn prevents bad deployments using "defective" build artifacts.

This specifically fixes a scenario where the requested package is a local workspace that requires some build transforms, as it is consumed directly from sources (e.g., a local workspace authored in TypeScript).