Update React to 18.2.0
Updates React to 18.2.0 across all packages. This is a devDependency upgrade for all packages. This is a peerDependency package upgrade only for the core packages that actually changed or incorporate React 18 features. For all other packages they will maintain a peer dep of >= 16.x until they otherwise to be bumped.