Add onRecoverableError handler to hydrateRoot
Instead of throwing an uncaught error on hydration mismatches in React 18, capture the uncaught using the `onRecoverableError` handler and re-fire it as a warning instead.

Going by our current classification of errors for Fusion, mismatches are not at the same level as other errors that are more severe and require immediate action. Developers should still strive to fix these though so we will continue logging them as warnings.