Fix React Router history bug where basename is not applied to initial server render
In the React Router 5 implementation, when the server history object was created, the prefix was prepended to the current URL: https://github.com/uber/fusionjs/pull/1409/files#diff-b9d2e3755ea5d8daffc083230c68bfcea742135f9fca8a8aa7f54d13467c6017L103

This behavior was incorrectly removed in the RR6 implementation. The history location should always have the basename/prefix as part of the URL since that's what React Router expects.

Also include a small fix to fusion-plugin-apollo where if we are in dev mode, to log out SSR errors the way we do in fusion-react. This is so the errors show up in dev.

Co-authored-by: shYkiSto <siarhei@uber.com>