Upgrade all projects to Jest 28.1.3
Upgrades Jest to 28.1.3 from the old 25.x version that was used prior.

Bulk of changes were because of Jest not allowing `async` and `done` to be used in the same test. So most were converted to drop the `async` and instead use promises instead while retaining the usage of `done`. In these cases, these tests still work because jest will wait for the allotted test time (default 5 seconds) for `done` to be called, allowing the test enough time to do its async behavior and assertions. If after that time `done` hasn't been called due to assertion failure or the test throwing, the test will fail.

There were a good number of tests that are `async` and also defined a done callback that didn't need it so I just removed the done callback.

There were a few changes of more importance:

1. [`jest-resolver-fusion-browser`](https://github.com/uber/fusionjs/pull/1530/files#diff-f281dca345b410750bc0fa0c1a758fadc8d741895a588d64f574ba01ed9b294a) was added for the few projects that needed this resolver in order for tests to work. The old solution was to have one shared package between server and browser but Jest removed support for reading the `browser: true` field in the jest configuration. So there wasn't an elegant way to share the custom resolver anymore and thus it had to split as a separate one.
2. [`fusion-core`](https://github.com/uber/fusionjs/pull/1530/files#diff-1d88b1de126f4a188756f3f0df97f912a4d4097aaaa2a9644cf2f53673cba512) required brand new snapshots since the old ones that defined the error stack available from tokens was changed. None of the code changed though so I chalked it up to Jest implementation details that resulted in changed snapshots.
3. [`fusion-plugin-error-handling`](https://github.com/uber/fusionjs/pull/1530/files#diff-7ce50697fec9acb0a77a870fb18300b53992e3de4deff5e26d73af67b1614260) was doing some Proxy magic that got bit by some new Jest function calls that happened on load. That was a pain to figure out and fix.
4. Some tests needed additional timeouts defined as the old ones were too short now.